Resolve upload order issue with by filter png delete on server by .png. Resolve cell annotation issue, but still size issues with annotating after tif.

+++ pythonclassifier/annotool.py
@@ -63,7 +63,7 @@ def train():
 	for the_file in os.listdir(folder):
 		file_path = os.path.join(folder, the_file)
 		try:
-			if os.path.isfile(file_path):
+			if os.path.isfile(file_path) and the_file.endswith(".png"):
 				os.unlink(file_path)
 				#elif os.path.isdir(file_path): shutil.rmtree(file_path)
 		except Exception as e:
